[MOV] rpc: move json2 controller in rpc module
The /json/2 controller doesn't depend on web, and is our main RPC entry
point. It is logical to place it inside the rpc module.

Used the opportunity to split the single controllers.py file, to expose
a jsonrpc.py and a xmlrpc.py file. With a bit of retro-compat in mind,
it is still possible to import the "RPC" controller and "_check_request"
via `odoo.addons.rpc.controllers`. The other symboles (primarely xmlrpc
ones) have been moved to `odoo.addons.rpc.controllers.xmlrpc`.
