+++ addons/rpc/controllers/json2.py
