+++ addons/web/controllers/__init__.py
@@ -7,7 +7,6 @@ from . import dataset
 from . import domain
 from . import export
 from . import json
-from . import json2
 from . import home
 from . import model
 from . import pivot
